requirements.txt
# Python requirements for git-circle
# ==================================
#
# This file is meant to be consumed by pip
#
#    pip install -r requirements.txt

requests
jinja
